refactor(worker): improve worker and more
